requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
decorator==4.4.2
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gensim==4.3.2
gTTS==2.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.25
langsmith==0.1.5
marshmallow==3.20.2
moviepy==1.0.3
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
passlib==1.7.4
pillow==10.2.0
priority==2.0.0
proglog==0.1.10
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
pyee==11.1.0
pymongo==4.6.2
pytesseract==0.3.10
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.12.0
smart-open==6.4.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
taskgroup==0.0.0a4
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
